+++ Changelog.md
@@ -10,6 +10,8 @@ Enhancements
 * Adds cucumber scenario showing how to invoke an anonymous controller's
   non-resourceful actions. (Paulo Luis Franchini Casaretto)
 * Null template handler is used when views are stubbed. (Daniel Schierbeck)
+* The generated `spec_helper.rb` in Rails 4 includes a check for pending
+  migrations. (Andy Lindeman)
 
 Bug fixes